Update init
I've added the live conversion for btrfs. For 256 RPi's the default memory split of 128 will give us to little memory to do a live conversion. This means we have to change the memory split for conversion and restore it afterwards. The actual booting from btrfs still doesn't work, so can you @mk01 implement and test that part? Not sure if the tune2fs is a proper check for an existing ext4 filesystem. We also need a check to confirm if the conversion failed. If so, we need to boot from the old ext4 filesystem.